feat(sql): make SQL the sole document query surface
Unify parsing, schemas, exact indexes, DDL/DML, joins, aggregates, and transactions behind one bounded typed runtime shared by database/sql and pgwire. Remove the competing JSON query grammars, add real pgx/libpq integration coverage, and harden cancellation, resource accounting, durability, and 32-bit hostile-input paths.

+++ store/durable/store_file_materialization_bench_test.go
@@ -23,7 +23,7 @@ func BenchmarkFileMaterializationDurableUpdate(b *testing.B) {
 	zoned := make([][]byte, 512)
 	var previousZoneCode uint32
 	for index := range zoned {
-		number := fmt.Sprintf("%012d", 100_000_000_000+index*1_000_000_000)
+		number := fmt.Sprintf("%012d", int64(100_000_000_000)+int64(index)*1_000_000_000)
 		code, ok := store.ZoneCodeNumber([]byte(number))
 		if !ok || index != 0 && code <= previousZoneCode {
 			b.Fatalf("zone benchmark code %q = %d after %d",
